MB-19948: Refactor CouchKVStore meta-data handling
Wrap the ep-engine metadata we store into couchstore in
a class to provide a much cleaner management interface.

This includes a fix for the orginal MB-19948 issue in that
we correctly read V1 metadata (the 18byte version written
by 3.x). The fix is kind of hidden in the refactor, but the
new code explictly looks at the size of metadata and has
construction paths for 16, 18 and 19 bytes.

Tests in cab19c11ea95df30a119196078f7bc8bff683692 lend
some confidence to this refactoring.

Change-Id: I7d2042af519819737a6e4d070fd8185c84778d6a
Reviewed-on: http://review.couchbase.org/64978
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>