MB-19833: Set a minimum boundary for config param bfilter_key_count
We can have a 'divide by zero' error if we have bfilter_key_count set
to 0 in configuration. Further it does not make sense to have the
param, which expects estimated key count per vbucket, to be set to 0.

This commit hence adds a validator to the config param to make sure
it is not below 1.

Change-Id: I5f361e87e053121b56a77593daab6c46990562dc
Reviewed-on: http://review.couchbase.org/64914
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>