MB-19957: Fix potential infinite hang
If an error occurs when we're trying to start the audit daemon
the destructor for the audit instance tries to join a thread with
a "random" id. Given that a cb_thread_t is an opaque structure
which we have no well defined values for we don't initialize it
to anything (We could do that on windows where we have
INVALID_HANDLE_VALUE, but there is no invalid values for
pthread_t). Instead we'll do a "poor" effort to work around this
by keeping the state in a boolean variable.

Change-Id: I91c854a4b03174926104853ed69b2921eef50475
Reviewed-on: http://review.couchbase.org/64973
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>