MB-19339: Refactor RBAC handling in memcached
memcached used to follow the least privilege principle by
allowing the "user" to grant access to individual commands.
As part of the task to integrate memcacheds RBAC into the
entire stack, we're refactoring the RBAC support to look
more like the model being used in ns_server.

We've defined a small subset of privileges and specify
which privileges each command require and perform that check
before each command is executed. Given that some commands
may produce different results depending on which privilege
the "user" have access to, the server API is also extended
to provide a method for the underlying engine to call to
check for privilege access:

    if (!sapi->cookie.check_privilege(cookie,
                                     Privilege::Read)) {
        return ENGINE_EACCESS;
    }

We've not yet decided on how memcached shall receive the
RBAC configuration defined by the user. There is currently
two alternatives:

1) Use the component in `ns_server`
      + The component already exists; only minor changes is needed
        to add support for our privileges
      + No need to standardize the "file format" for the RBAC data
      + Only one implementation of the access evaluation
            + The same bugs exists everywhere
      - memcached depends on the availability of another component
            - what to do if it refuse to answer on the port
            - what to do if we're having "network" failures
            - what to do if we're failing to create a socket to perform
              the rest call (running out of file descriptors)
      - We need to create a mock of the server for our unit testing
            - need to implement enough failure scenarios to ensure we handle
              all kinds of problems
      - Is it fast engough, and does it scale?
      - We need to add support for HTTP in memcached
      - Cache invalidation
            - We would have to generate the privilege set every time
              the user selects a bucket (or performs a new auth), but
              we would have no clue when to invalidate this. We _could_
              add a TTL for the privilege set for let's say 1h to avoid
              having to reboot all nodes to kick out all users which no
              longer have access to the privileges.

2) Let `memcached` implement the access control
      + No external process dependencies affecting the availability or
        error situations at runtime
      - Need a fixed file format between `ns_server` and `memcached` and
        a way to signal memcached that the files changed.
            - Need to implement parser in multiple components
            + Need to create test suite to ensure that all components
              produce the same result of the given input
      + Extremely fast and scalable. Everything is lookup of internal
        datastructures in a "copy on write" mode. Lock only held in
        order to create a shared pointer to the datastructures. Cache
        invalidation performed by looking at two atomic variables.
      - No need for a mock server, all tests may be performed with
        real configuration data.
      + Simpler logic in `memcached`
            + No need to suspend a connection in the middle of the AUTH /
              Select-Bucket phase in order to request the privilege
              set from `ns_server`

Given the above we can't currently provide any useful implementation
with a fine grained RBAC control, so this patch grants connections
authenticated as _admin full access to the system. All other
connections will get access similar to what they used to get before
RBAC was added to memcached.

The unit tests needs to be updated to make sure they use the new
privilege infrastructure when we have decided if we want to use
an internal component or an external component.

Change-Id: I4154053a2394ac836f34593b6f6067b0a46f5d71
Reviewed-on: http://review.couchbase.org/65148
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>