MB-16661: Server side moxi should be removed
As part of MB-16661 server side moxi should be removed. This is a
temporary patch that adds a binary named moxi to the system so
that ns_server can start and control moxi the way it used to
until the ns_server team finds time to prioritize the bug
report (this saves us from maintaining the dependencies, source
and tests in moxi while we're waiting).

Change-Id: Ifc7471b1c476f982fc5482a2db7ab70b0f73ce57
Reviewed-on: http://review.couchbase.org/65334
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>