MB-20061: Improved futureQueue heap property fix
MB-9986 found that we break the heap property of the underlying
container of the std::priority_queue used in scheduling. This is
because we use the wakeTime to sort the queue, but then later
change the value of a task which is still in the queue, thus
order is broken. MB-9986 emptied the futureQueue and rebuilt it.

In this fix we utilise the C++ standard which shows that we can
access the container of the priority_queue. Thus we can sub-class
it and create a method that allows us to update a tasks time and
maintain the heap order.

Change-Id: I238d36ea684d59ef06326183fa1f16c04f8d29ad
Reviewed-on: http://review.couchbase.org/65394
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>