MB-19981 Call registerEvent as opposed to updateEvent
If a bucket delete is attempted while there is connection in the
conn_nread state.  And that connection is currently blocked waiting
for a response from the server; the connection will not have an event
registered in libevent.  Therefore a call to updateEvent will fail.

This patch changes the code in signalIfIdle to use registerEvent
as opposed to updateEvent in such cases.

A corresponding test is also added.  Before the fix, if the
event_active function call is removed from the signalIfIdle function
the test will hang.  The reason the test works with the event_active
function call in place is that the event_active function
can be invoked regardless of whether the event is registered
(i.e. in a pending state) or not.

Change-Id: Ia205180f97ad58b772efc99b09194f6c640ef462
Reviewed-on: http://review.couchbase.org/65538
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>