MB-19955: Reduce NOOP interval to one second
ns_server plan to spy on dcp traffic to determine whether a
node is healthy.  Therefore on an idle node they need to
ensure a DCP NOOP message is sent at a one second interval.

This patch reduces the NOOP interval to one second.
It also modifies the ConnManager task to invoke the
manageConnections function once a second.

The connection timeout is changed from being twice the NOOP
interval to being a separate dedicated configuration option.

The decision of whether to disconnect is moved outside the
maybeSendNoop function into a dedicated maybeDisconnect
function.

Change-Id: I1b47e46fcd84a8d7f6a2f885b0120863d031b251
Reviewed-on: http://review.couchbase.org/65927
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>