Perform an explicit commit after call to flushVBucket
Although a flushVBucket writes the item to the underlying store,
the item is not marked clean until an explicit commit is called.
If the underlying store is couchstore, a commit is called with
a flushVBucket but in the case of forestdb, a commit is not
always called, hence call an explicit commit.

Change-Id: I2feee981f68b9302b9c65d002dae55a19d1da4c2
Reviewed-on: http://review.couchbase.org/65670
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>