Refactor: Let the core implement append/prepend
This was initially added to the underlying engines so that they
could implement append/prepend more efficiently by using different
chunks of memory, but we never ended up doing that. Now that
we are going to add the XATTRs we would like to avoid having
all underlying engines to have code to handle that so we're
just moving the already existing append/prepend logic up from
the clients.

Change-Id: If75d6b52cb09e392a2ffa09fc10991e40b6f5bdf
Reviewed-on: http://review.couchbase.org/65776
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>