MB-20203: Allow document creation with subdoc
This allows the document to be created if it does not exist, and if a
new flag (SUBDOC_FLAG_MKDOC, implies SUBDOC_FLAG_MKDIR_P). This can be
used to add items to a queue or create a counter, for example.

This commit depends on a related change within subjson.

Change-Id: If78ce56d6eba89e9f540cd77f11a0bdfd5292393
Reviewed-on: http://review.couchbase.org/65962
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>