MB-20201: Create a new handle only if needed
When we try to fetch the associated KV store handle for
a specific vbucket, we always create one if one if not
found. For some cases like delete vbucket and fetch,
we are just trying to close the existing handle and
we don't necessarily have to create a new one.

Change-Id: I98195801c4cfb6e8cfecda124198a23b6915ada6
Reviewed-on: http://review.couchbase.org/66553
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>