MB-20357 Modernize ioctl handling
This patch modernizes and C++ifies the ioctl handling to make it
easier to add new ioctl options and to make it possible to return
non-integers for an ioctl get operation.

This patch also fixes a bug in the tcmalloc aggressive memory
decommit test on Windows which doesn't support the printf '%zd'
flag on versions of MSVC before 2015. It also temporarily removes
the `trace.connection` option which relied on a prefixing scheme
in lieu of a replacement system to be added in a subsequent
patch.

Change-Id: I21d7bc15723d87a704d3d0641c39dc3cbee844fd
Reviewed-on: http://review.couchbase.org/66865
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>