MB-20358 Add 'trace.dump' IOCTL command
This adds a basic 'dump' implementation for dumping a trace across
the network. This currently has the downside of dumping it in one
go (Which might be *A LOT*) and is intended as a developer-mode MVP
with more robust behaviour to come in future.

Usage:

    ./mcctl get trace.dump

Will output the trace to stdout as one continuous JSON object. It
will also stop tracing if it is currently running.

Change-Id: I356f7cb98af1ed840108ad2be9029a0196a86647
Reviewed-on: http://review.couchbase.org/67033
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>