Estimating cummulative flushtimes & committimes
Calculate the cumulativeFlushTime and cumulativeCommitTime
with the same granularity of milli-seconds.

Use gethrtime() to set the flush_start time, as gethrtime()
is used to set the end, which is used to estimate transaction
time for an item. Also the same start and end times can be
used to calculate the cumulativeFlushTime as well.

Change-Id: Iea77e49715303afbdbbd1fd7f9d8eb8f10155843
Reviewed-on: http://review.couchbase.org/67010
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>