MB-20586: darwin_zone: Use correct handle for zone_calloc
Fix a typo in our darwin zone allocator / shim - use the underlying
zone handle when calling into default_zone->calloc().

This appears to be a benign issue when the default zone is jemalloc
(as is the standard deployment), but can causes segfaults when using
the system zone allocator.

Change-Id: I4d525a5482766b5a6f5b272cf6f994d00dc9cc35
Reviewed-on: http://review.couchbase.org/67086
Tested-by: buildbot <build@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>