MB-20735: Fix incorrect waketime for executor threads
When a thread is done with a task from a particular queue,
it calls doneWork() which was resetting its current Queue type
resulting in subsequent reschedule to pick infinity as
earliest thread waketime and end up sleeping longer when the
readyQueue is empty.
Simplify the reschedule logic as part of this fix.
+Add unit test for validation.

Change-Id: Ifdad1d21e624982460046393d4c377247da90735
Reviewed-on: http://review.couchbase.org/67112
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: buildbot <build@couchbase.com>