MB-20586: Add plugin test to memory_tracking_test, remove memalign test
Expand the memory_tracking_test to add a set of tests which verify that
memory allocated in a dlopen'd plugin is correctly tracked.

This is in preparation for the cbmalloc-based tracking where we do not
implicilty get all calls to malloc() (in any library) tracked.

Also remove the memalign tests - we don't actually use memalign in our
codebase, and hence having to support it in the cbmalloc API just for
testcode is unnecessary.

Change-Id: I59a0570c1fb64059cb94dc68f1ded366df2385fc
Reviewed-on: http://review.couchbase.org/67303
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: buildbot <build@couchbase.com>