MB-20794: Join parent monitor thread before shutdown
Currently the parent monitor thread in memcached is created detached,
this means that we do not get a chance to clean it up before
shutdown. Historically hasn't really mattered, however with the
introduction of phosphor event tracing and registering all thread with
Phosphor we allocate memory (for the thread's name) as part of thread
creation.

To ensure this memory is not leaked on shutdown (creating Valgrind
warnings), do not dettach parent monitor and join it before shutdown.

Change-Id: Ic9f5f6792b6cc6667fa64f90cacbc9bb851b042a
Reviewed-on: http://review.couchbase.org/67387
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>