MB-20834: Use get_available_cpu_count() for #workers calculation
Use the new Couchbase::get_available_cpu_count() when calculating how
many worker threads to create. get_available_cpu_count() accounts for
the number of logical cores a process is allowed to run on, which is
the mechanism cgroups (and hence Docker) uses to limit the CPUs
available to a container.

This fixes the problem of us creating "too many" worker threads under
Docker containers which have the --cpuset-cpus option set.

Change-Id: I3e3b91eecc51aea298ae9aceb9e8c6d3f16b7612
Reviewed-on: http://review.couchbase.org/67489
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>