MB-20586: memcached_testapp: Fix leaks with OpenSSL
If we run GoogleTest in a configuration which doesn't include any
tests, then it never calls the SetUp and TearDown() methods of any
testing::Environment objects. This is problematic as we rely on
shutdown_openssl() being called to free all the memory allocated by
OpenSSL - including memory which is allocated on shared library
open. In other words even if we never call initialize_openssl we need
to call shutdown_openssl().

To fix this move the init and shutdown OpenSSL calls to the
constructor / destructor, which /are/ always called.

Change-Id: I55696f1198bd81bcb7d3a14f51092dc91022addf
Reviewed-on: http://review.couchbase.org/67571
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>