Remove default parameter from evictKey
The EventuallyPersistentStore method evictKey has
a parameter called force, which is defaulted to
false.  The parameter is used to specify that we
want to force the item to be deleted.

However the evictKey method is never passed
a force parameter and therefore the default
of false is always used.

This patch removes the unrequired parameter from
the evictKey definition and simpifies the body of
the method.

Change-Id: I8cc1ba49ca0331835a9d797b140d6c3666abf8a8
Reviewed-on: http://review.couchbase.org/67594
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>