Remove default parameter from getMetaData
The EventuallyPersistentStore method getMetaData has a parameter
called trackReference, which is defaulted to false.  The parameter
is used to specify if we want to set the NRU bit for the item.

However the getMetaData method only uses the trackReference default
of false.

This patch removes the unrequired parameter from the getMetaData
definition and simpifies the body of the method.

Change-Id: I8f2b9ff911ce188d6d8b5b087e378b47a4cdaa8e
Reviewed-on: http://review.couchbase.org/67628
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>