Add support for ldap via saslauthd
ns_server allows for authentication via LDAP by utilizing
saslauthd. This patch adds the same functionality to memcached.

In order for memcached to enable saslauthd, the environment
variable CBAUTH_SOCKPATH must be set to the unix domain socket
to communicate to saslauthd.

The user needs to be defined as external in the cbsasl password
database for it to be looked up in LDAP.

Change-Id: Id634977018b52c777bf71b6c69ab53859f947140
Reviewed-on: http://review.couchbase.org/67668
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>