Treat all unknown users as external users
This means that all ns_server needs to do is to update
the privilege map (to be added as part of the RBAC work).

By doing so you can authenticate to the cluster, but you'll be
associated with the no-bucket and disconnected until you
run either select bucket (which you currently can't).

Change-Id: I381dc0559545c29dbd19620c2a9eb14eebd06e66
Reviewed-on: http://review.couchbase.org/67735
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>