MB-20852: ep_unit_tests_main: Show DEBUG logs with -v
Previously we would only print INFO level log messages when the -v
flag was given, even though the code claimed to have enabled DEBUG
level.

This was because we didn't get the log level early enough - the
extension defaults to INFO, and will propogate that down to ep_engine
(overwriting what we specify). Fix by setting the log level in the
server API to debug.

Change-Id: I8a192169721fad631965600b69975d882b6221f8
Reviewed-on: http://review.couchbase.org/67888
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>