Engine_testapp: Phosphor to dump out log before unloading engine
Phosphor has tracepoint_infos which hold static data for a given
tracepoint.

The tracepoint_info is declared statically and therefore can be
allocated within the bss of the engine library.  As the
tracepoint_info is accessed when the logs are written out, it is
important that this process happens prior to unloading the engine.

This patch ensures that PHOSPHOR_INSTANCE.stop is called prior to
unloading the engine.

Change-Id: Ib240becefb9f5f831365e0937aa1ad92cf09a24c
Reviewed-on: http://review.couchbase.org/67916
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: buildbot <build@couchbase.com>