Add support for enabling xattr support
The DCP producer must be told explicitly to send XATTRs to
the consumer to support backwards compatibility.

The DCP client in ep-engine don't need support for this
control message as it is a message sent from the client to
the server. By default the consumer in ep-engine expects
ns_server to enable xattr support.

Change-Id: I71e6b570058314f663f1c9d9427740c46754c8ea
Reviewed-on: http://review.couchbase.org/68625
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>