Make the /all/ list TAP specific in Connmap
Connmap maintains a map of cookies to connections, as well as a list
of all connections.

In TAP the all list may not always match the entries in the map
because dead connections are removed from the map, but remain in the
all list and cleaned-up at a later time.

In DCP we maintain a separate deadConnections list and therefore map_
is always equivalent to the all list and therefore we need only refer
to the map.

This patch removes the use of the all list from DCP.

Change-Id: I7e892599168e0b4db837a53ed5d24f44af702aab
Reviewed-on: http://review.couchbase.org/68229
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>