Refactor: Introduce struct to pass stats cleanly
In ep_perfsuite we want to record timings for each of the couchbase
stats.  We pass around vectors containing the timings for each of
the individual stats.  However, as there are a large number of stats
this patch introduces a StatOperationTimings struct which is used as
a container for all of the stats timings.

Change-Id: I403c0fb8b9a1da7c630a11e714cbbb7ca0784b1b
Reviewed-on: http://review.couchbase.org/68277
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>