Use == and compare instead of strncmp in getStats
Motivation for this change was that we had a bug where we attempting
to string compare for 'hash' however, we passed in 3 as opposed to 4
into strncmp - and so only compared 'has'.

This patch replaces the use of strncmp with == and string::compare.
It also clean-ups some of the code, whilst maintaing the same logic.

Change-Id: I0b8fb8169d84698e293779c8e7d2bb6e75dc9686
Reviewed-on: http://review.couchbase.org/68473
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>