MB-21401: bump the "slow command" limit for delete bucket
A delete bucket command needs to disconnect all connected
clients (which may take some time as they need to complete
the current operation); shut down the internal tasks etc.

This means that on a bucket in production it is more or
less expected for the operation to take more than 500ms.

Change-Id: Icab1158def2d89c68b9cf6c4072fb8823eef68be
Reviewed-on: http://review.couchbase.org/68849
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>