Revert "Add support for enabling xattr support"
This reverts commit 45ab791d813c01337326c098eb5a9f6c24316fab.

We're going to encode the (presence and length) of xattrs as
part of the flex meta field. By doing so we should enable the
knowledge of them by using HELO with XATTR support (this is the
same way we'll be enabling collection aware flex meta to be
encoded in GetWithMeta and SetWithMeta).

This means that we no longer need the control message.

Change-Id: I3d28e720a8eb0b673b1eec0453599789d14766b8
Reviewed-on: http://review.couchbase.org/68917
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>